+++ runway/templates/cdk-py/app.py
@@ -1,0 +1,42 @@
+"""Sample app CDK module."""
+from aws_cdk import aws_lambda, cdk
+
+
+class LambdaStack(cdk.Stack):
+    """CFN stack class."""
+
+    def __init__(self, app: cdk.App, id: str) -> None:
+        """Create LambdaStack."""
+        super().__init__(app, id)
+
+        with open('lambda-index.py', encoding='utf8') as stream:
+            handler_code = stream.read()
+
+        lambdaFn = aws_lambda.Function(
+            self,
+            'MyFunction',
+            code=aws_lambda.InlineCode(handler_code),
+            handler='index.main',
+            timeout=300,
+            runtime=aws_lambda.Runtime.PYTHON37,
+        )
+
+
+class App(cdk.App):
+    """CDK app class."""
+
+    def __init__(self) -> None:
+        """Create App."""
+        super().__init__()
+
+        LambdaStack(self, 'LambdaStack')
+
+
+def main() -> None:
+    """Handle cli invocation."""
+    app = App()
+    app.run()
+
+
+if __name__ == "__main__":
+    main()