+++ runway/templates/cdk-py/lambda-index.py
@@ -1,0 +1,6 @@
+"""Lambda function module."""
+
+
+def handler(event, context):
+    """Handle function invocations."""
+    print('Hello world!')